// Back/src/enum.ts
export enum TypeContrat {
  CDI = "CDI",
  CDD = "CDD",
  STAGE = "stage",
}

export enum Role {
  Admin = "Admin",
  RH = "RH",
  Superviseur = "Superviseur",
  Employe = "Employe"
}

export enum TypePointage {
  ENTREE = "entree",
  SORTIE = "sortie"
}

export enum ModePointage {
  BIO = "bio",
  BADGE = "badge",
  MANUEL = "manuel"
}

export enum StatutPointage {
  NORMAL = "normal",
  RETARD = "retard",
  AVANCE = "avance"
}

export enum StatutTimeSheet {
  ATTENTE = "attente",
  VALIDE = "valide",
  REJETE = "rejete"
}

export enum TypeMenu {
  PDEJ = "pdej",
  DEJ = "dej",
  REPAS = "repas",
  COLLATION = "collation"
}

export enum horaire{
  jour= "jour",
  nuit= "nuit",
  repos= "repos"

}